feat(register edit book page): create desktop layout
