+++ lib/pages/register_edit_book_page/commands/register_edit_book_page_desktop_command.dart
@@ -1,3 +1,3 @@
 import 'package:volume_vault/pages/register_edit_book_page/commands/register_edit_book_page_strategy.dart';
 
-class RegisterEditBookPageCommand extends RegisterEditBookPageStrategy {}
+class RegisterEditBookPageDesktopCommand extends RegisterEditBookPageStrategy {}
