+++ lib/pages/register_edit_book_page/commands/register_edit_book_page_mobile_command.dart
@@ -1,0 +1,3 @@
+import 'package:volume_vault/pages/register_edit_book_page/commands/register_edit_book_page_strategy.dart';
+
+class RegisterEditBookPageMobileCommand extends RegisterEditBookPageStrategy {}
